+++ tests/gen_prompt/prompt_corpus_node/prompt_corpus_node_parse_test.py
@@ -1,13 +1,13 @@
 """
-test .parse() and instance creation for ``PromptCorpusNode``
+test .parse() and instance creation for ``class PromptCorpusNode``
 """
 
 from kaye.gen_prompt import PromptCorpusNode
-from tests.gen_prompt.pcn.pcn_testee_corpus import (
+from tests.gen_prompt.prompt_corpus_node.testees import (
     PROMPT1,
     PROMPT2,
     PROMPT3,
-    PROMPT_EMPTYLINES,
+    PROMPT_EMPTY_LINES,
 )
 
 
@@ -285,7 +285,7 @@ class TestParse3:  # test using PROMPT3
 class TestEmptyLine:  # source material contains various empty lines
 
     def test_root(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
 
         assert tree.depth == 0
         assert tree.parent is None
@@ -293,7 +293,7 @@ class TestEmptyLine:  # source material contains various empty lines
         assert tree.content == []
 
     def test_project(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
         project = tree.children[0]
 
         assert project.name == "Project Title"
@@ -303,7 +303,7 @@ class TestEmptyLine:  # source material contains various empty lines
         assert project.content == []
 
     def test_description(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
         project = tree.children[0]
         sub = project.children[0]
 
@@ -312,11 +312,11 @@ class TestEmptyLine:  # source material contains various empty lines
         assert sub.parent is project
         assert len(sub.children) == 0
         assert sub.content == [
-            "A brief overview of the project, its purpose, and goals."
+            "A brief overview of the project, its purpose, and goals.",
         ]
 
     def test_install(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
         project = tree.children[0]
         sub = project.children[1]
 
@@ -331,7 +331,7 @@ class TestEmptyLine:  # source material contains various empty lines
         ]
 
     def test_usage1(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
         project = tree.children[0]
         sub = project.children[2]
 
@@ -340,11 +340,11 @@ class TestEmptyLine:  # source material contains various empty lines
         assert sub.parent is project
         assert len(sub.children) == 0
         assert sub.content == [
-            "Provide instructions on how to use the application."
+            "Provide instructions on how to use the application.",
         ]
 
     def test_usage2(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
         project = tree.children[0]
         sub = project.children[3]
 
@@ -359,7 +359,7 @@ class TestEmptyLine:  # source material contains various empty lines
         ]
 
     def test_license(self):
-        tree = PromptCorpusNode.parse(PROMPT_EMPTYLINES)
+        tree = PromptCorpusNode.parse(PROMPT_EMPTY_LINES)
         project = tree.children[0]
         sub = project.children[4]
 
@@ -368,7 +368,7 @@ class TestEmptyLine:  # source material contains various empty lines
         assert sub.parent is project
         assert len(sub.children) == 0
         assert sub.content == [
-            "This project is licensed under the MIT License."
+            "This project is licensed under the MIT License.",
         ]